Refactor: 분리된 파일 동일
